backend/requirements.txt

fastapi==0.103.1
uvicorn[standard]==0.22.0
motor==3.1.1
pymongo==4.3.3
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic>=2.0
requests== 2.32.3


